Testing that we do not update the TTL of the session when we get/set it.
